requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
openai==1.58.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-dotenv==1.0.1
apscheduler==3.11.0
jinja2==3.1.4
python-multipart==0.0.20
